# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Absensi",
        "description": "API Aplikasi absensi berbasis web (e-absensi) online untuk memantau kinerja karyawan jarak jauh, dan pekerja work from home (WFH).",
        "termsOfService": "http://127.0.0.1:8000/terms/",
        "contact": {
            "email": "bandungcoders@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Phone number user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get auth user",
                "description": "",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh a token",
                "description": "",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "description": "",
                "operationId": "password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "description": "Old password user",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "new_confirm_password": {
                                        "description": "password confirm",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot password",
                "description": "",
                "operationId": "forgot",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email user",
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validation one time password",
                "description": "",
                "operationId": "otp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "otp": {
                                        "description": "OTP",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password",
                "description": "",
                "operationId": "reset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "otp": {
                                        "description": "OTP",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "new_confirm_password": {
                                        "description": "Password Confirm",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
...
    },
...
}